version: move version library to its own folder
This is to place nicer with the expectations of Bazel.
